Create code_2
The function first checks if the denominator is zero, and if it is, it throws an error message. If the denominator is not zero, the function returns the result of dividing the numerator by the denominator.

You can then call this function by passing in your desired numerator and denominator as arguments.